requirements.txt
addict==2.4.0
aiohttp==3.8.6
aiosignal==1.3.1
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.1
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
brotlipy==0.7.0
cachetools==5.3.2
certifi @ file:///C:/b/abs_36eb5mzhph/croot/certifi_1690232276943/work/certifi
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
chumpy==0.70
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
coverage==7.3.0
crcmod==1.7
cryptography @ file:///C:/b/abs_13590mi9q9/croot/cryptography_1689373706078/work
cycler==0.11.0
Cython==3.0.2
decord==0.6.0
docker-pycreds==0.4.0
einops==0.7.0
exceptiongroup==1.1.3
fastapi==0.104.0
ffmpeg==1.4
filelock @ file:///C:/b/abs_c7yrhs9uz2/croot/filelock_1672387617533/work
filterpy==1.4.5
flake8==6.1.0
fonttools==4.42.1
frozenlist==1.4.0
fvcore==0.1.5.post20221221
gitdb==4.0.10
GitPython==3.1.35
google-api-core==2.12.0
google-auth==2.23.3
google-cloud-core==2.3.3
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
imageio==2.31.3
importlib-metadata==6.8.0
importlib-resources==6.0.1
iniconfig==2.0.0
interrogate==1.5.0
iopath==0.1.10
isort==4.3.21
Jinja2 @ file:///C:/b/abs_7cdis66kl9/croot/jinja2_1666908141852/work
jmespath==0.10.0
joblib==1.3.2
json-tricks==3.17.3
kiwisolver==1.4.5
lazy_loader==0.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/ci/markupsafe_1654489871526/work
matplotlib==3.7.2
mccabe==0.7.0
mdurl==0.1.2
mkl-fft==1.3.6
mkl-random @ file:///C:/Users/dev-admin/mkl/mkl_random_1682977971003/work
mkl-service==2.4.0
mmaction2==1.1.0
mmcv==2.0.1
mmcv-full==1.7.0
mmdet==3.1.0
mmengine==0.8.4
-e git+https://github.com/byunnn/mmpose.git@e8f259bcd25d13b043bddf8c2f37990c457bfbb2#egg=mmpose
model-index==0.1.11
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
multidict==6.0.4
munch==4.0.0
munkres==1.1.4
networkx @ file:///C:/b/abs_e6gi1go5op/croot/networkx_1690562046966/work
numpy @ file:///C:/Users/dev-admin/mkl/numpy_and_numpy_base_1682982345978/work
opencv-contrib-python==4.8.1.78
opencv-python==4.8.0.76
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.24
ordered-set==4.1.0
oss2==2.17.0
packaging==23.1
pandas==2.0.3
parameterized==0.9.0
pathtools==0.1.2
Pillow==9.4.0
platformdirs==3.10.0
pluggy==1.3.0
portalocker==2.7.0
protobuf==4.24.2
psutil==5.9.5
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycodestyle==2.11.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.18.0
pydantic==2.4.2
pydantic_core==2.10.1
pyflakes==3.1.0
Pygments==2.16.1
pyOpenSSL @ file:///C:/b/abs_08f38zyck4/croot/pyopenssl_1690225407403/work
pyparsing==3.0.9
PySocks @ file:///C:/ci/pysocks_1605287845585/work
pytest==7.4.1
pytest-runner==6.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytorch-fid==0.3.0
pytz==2023.3.post1
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0.1
regex==2023.8.8
requests==2.28.2
rich==13.4.2
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.13.0
sentry-sdk==1.30.0
setproctitle==1.3.2
shapely==2.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.27.0
sympy @ file:///C:/b/abs_95fbf1z7n6/croot/sympy_1668202411612/work
tabulate==0.9.0
termcolor==2.3.0
terminaltables==3.1.10
threadpoolctl==3.2.0
tifffile==2023.7.10
toml==0.10.2
tomli==2.0.1
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torcheval==0.0.7
torchsummary==1.5.1
torchvision==0.15.2+cu117
tqdm==4.65.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3 @ file:///C:/b/abs_889_loyqv4/croot/urllib3_1686163174463/work
uvicorn==0.23.2
wandb==0.15.10
watchfiles==0.21.0
websockets==12.0
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
xdoctest==1.1.1
xtcocotools==1.14
yacs==0.1.8
yapf==0.40.1
yarl==1.9.2
zipp==3.16.2
